Don't show artifact transformation headers
The headers for artifact transforms should not show up in the build
log if there is no output from the transformations.
That makes the console log much cleaner, especially for the plain and
verbose console.

The logging for artifact transformations still shows up in build scan
logs or if the transformation produced some output.